mysite/requirements.txt
Django==2.2
mysqlclient==1.4.6
djangorestframework==3.11.0
google-api-python-client
jsonfield==3.1.0

